PARTIAL WIN32 PORT + Paso, ext_math and mmio libraries build using scons in top level directory + Scons build system currently hard coded against RW's third-party library paths. + Still lots TODO to complete port and build system

+++ escript/test/Data/DataTestCase.cpp
@@ -406,7 +406,7 @@ void DataTestCase::testOperations() {
   Data result(base.powD(power));
   for (int i=0;i<shape[0];i++) {
     for (int j=0;j<shape[1];j++) {
-      assert(result.getPointDataView()(i,j) == pow(dataView.index(i,j),3.0));
+      assert(result.getPointDataView()(i,j) == pow((double)dataView.index(i,j),3.0));
     }
   }
